Moved storage writer open and close out of extraction engine
