+++ plaso/cli/log2timeline_tool.py
@@ -6,8 +6,6 @@ import sys
 import time
 import textwrap
 
-from dfvfs.lib import definitions as dfvfs_definitions
-
 import plaso
 
 # The following import makes sure the output modules are registered.
@@ -16,18 +14,13 @@ from plaso import output  # pylint: disable=unused-import
 from plaso.analyzers.hashers import manager as hashers_manager
 from plaso.cli import extraction_tool
 from plaso.cli import logger
-from plaso.cli import status_view
 from plaso.cli import tools
 from plaso.cli import views
 from plaso.cli.helpers import manager as helpers_manager
-from plaso.engine import engine
-from plaso.engine import single_process as single_extraction_engine
 from plaso.lib import definitions
 from plaso.lib import errors
 from plaso.lib import loggers
-from plaso.multi_processing import extraction_engine as multi_extraction_engine
 from plaso.parsers import manager as parsers_manager
-from plaso.storage import factory as storage_factory
 
 
 class Log2TimelineTool(extraction_tool.ExtractionTool):
@@ -71,11 +64,6 @@ class Log2TimelineTool(extraction_tool.ExtractionTool):
       'And that is how you build a timeline using log2timeline...',
       '']))
 
-  _SOURCE_TYPES_TO_PREPROCESS = frozenset([
-      dfvfs_definitions.SOURCE_TYPE_DIRECTORY,
-      dfvfs_definitions.SOURCE_TYPE_STORAGE_MEDIA_DEVICE,
-      dfvfs_definitions.SOURCE_TYPE_STORAGE_MEDIA_IMAGE])
-
   def __init__(self, input_reader=None, output_writer=None):
     """Initializes a log2timeline CLI tool.
 
@@ -87,12 +75,7 @@ class Log2TimelineTool(extraction_tool.ExtractionTool):
     """
     super(Log2TimelineTool, self).__init__(
         input_reader=input_reader, output_writer=output_writer)
-    self._command_line_arguments = None
-    self._enable_sigsegv_handler = False
-    self._number_of_extraction_workers = 0
     self._storage_serializer_format = definitions.SERIALIZER_FORMAT_JSON
-    self._status_view = status_view.StatusView(self._output_writer, self.NAME)
-    self._status_view_mode = status_view.StatusView.MODE_WINDOW
     self._stdout_output_writer = isinstance(
         self._output_writer, tools.StdoutOutputWriter)
 
@@ -382,125 +365,6 @@ class Log2TimelineTool(extraction_tool.ExtractionTool):
     self._enable_sigsegv_handler = getattr(options, 'sigsegv_handler', False)
 
     self._EnforceProcessMemoryLimit(self._process_memory_limit)
-
-  def ExtractEventsFromSources(self):
-    """Processes the sources and extracts events.
-
-    Raises:
-      BadConfigOption: if the storage file path is invalid or the storage
-          format not supported or an invalid collection filter was specified.
-      SourceScannerError: if the source scanner could not find a supported
-          file system.
-      UserAbort: if the user initiated an abort.
-    """
-    self._CheckStorageFile(self._storage_file_path, warn_about_existing=True)
-
-    self.ScanSource(self._source_path)
-
-    is_archive = False
-    if self._source_type == dfvfs_definitions.SOURCE_TYPE_FILE:
-      is_archive = self._IsArchiveFile(self._source_path_specs[0])
-      if is_archive:
-        self._source_type = definitions.SOURCE_TYPE_ARCHIVE
-
-    self._status_view.SetMode(self._status_view_mode)
-    self._status_view.SetSourceInformation(
-        self._source_path, self._source_type,
-        artifact_filters=self._artifact_filters,
-        filter_file=self._filter_file)
-
-    status_update_callback = (
-        self._status_view.GetExtractionStatusUpdateCallback())
-
-    self._output_writer.Write('\n')
-    self._status_view.PrintExtractionStatusHeader(None)
-    self._output_writer.Write('Processing started.\n')
-
-    session = engine.BaseEngine.CreateSession(
-        artifact_filter_names=self._artifact_filters,
-        command_line_arguments=self._command_line_arguments,
-        debug_mode=self._debug_mode,
-        filter_file_path=self._filter_file,
-        preferred_encoding=self.preferred_encoding,
-        preferred_time_zone=self._preferred_time_zone,
-        preferred_year=self._preferred_year,
-        text_prepend=self._text_prepend)
-
-    storage_writer = storage_factory.StorageFactory.CreateStorageWriter(
-        self._storage_format, session, self._storage_file_path)
-    if not storage_writer:
-      raise errors.BadConfigOption('Unsupported storage format: {0:s}'.format(
-          self._storage_format))
-
-    single_process_mode = self._single_process_mode
-    if self._source_type == dfvfs_definitions.SOURCE_TYPE_FILE:
-      if not self._process_archives or not is_archive:
-        single_process_mode = True
-
-    if single_process_mode:
-      extraction_engine = single_extraction_engine.SingleProcessEngine()
-    else:
-      extraction_engine = multi_extraction_engine.ExtractionMultiProcessEngine(
-          number_of_worker_processes=self._number_of_extraction_workers,
-          worker_memory_limit=self._worker_memory_limit,
-          worker_timeout=self._worker_timeout)
-
-    # If the source is a storage media image or device, or directory
-    # run pre-processing.
-    if self._source_type in self._SOURCE_TYPES_TO_PREPROCESS:
-      self._PreprocessSources(extraction_engine, storage_writer)
-
-    configuration = self._CreateProcessingConfiguration(
-        extraction_engine.knowledge_base)
-
-    session.enabled_parser_names = (
-        configuration.parser_filter_expression.split(','))
-    session.parser_filter_expression = self._parser_filter_expression
-
-    self._SetExtractionPreferredTimeZone(extraction_engine.knowledge_base)
-
-    # TODO: set mount path in knowledge base with
-    # extraction_engine.knowledge_base.SetMountPath()
-    extraction_engine.knowledge_base.SetTextPrepend(self._text_prepend)
-
-    try:
-      extraction_engine.BuildCollectionFilters(
-          self._artifact_definitions_path, self._custom_artifacts_path,
-          extraction_engine.knowledge_base, self._artifact_filters,
-          self._filter_file)
-    except errors.InvalidFilter as exception:
-      raise errors.BadConfigOption(
-          'Unable to build collection filters with error: {0!s}'.format(
-              exception))
-
-    processing_status = None
-    if single_process_mode:
-      force_parser = False
-      number_of_parsers = len(configuration.parser_filter_expression.split(','))
-      if (self._source_type == dfvfs_definitions.SOURCE_TYPE_FILE and
-          not is_archive and number_of_parsers == 1):
-        force_parser = True
-
-      logger.debug('Starting extraction in single process mode.')
-
-      processing_status = extraction_engine.ProcessSources(
-          session, self._source_path_specs, storage_writer,
-          self._resolver_context, configuration, force_parser=force_parser,
-          status_update_callback=status_update_callback)
-
-    else:
-      logger.debug('Starting extraction in multi process mode.')
-
-      # The following overrides are needed because pylint 2.6.0 gets confused
-      # about which ProcessSources to check against.
-      # pylint: disable=no-value-for-parameter,unexpected-keyword-arg
-      processing_status = extraction_engine.ProcessSources(
-          session, self._source_path_specs, storage_writer,
-          configuration, enable_sigsegv_handler=self._enable_sigsegv_handler,
-          status_update_callback=status_update_callback,
-          storage_file_path=self._storage_file_path)
-
-    self._status_view.PrintExtractionSummary(processing_status)
 
   def ShowInfo(self):
     """Shows information about available hashers, parsers, plugins, etc."""
