+++ plaso/cli/psteal_tool.py
@@ -8,24 +8,17 @@ import textwrap
 
 from dfdatetime import posix_time as dfdatetime_posix_time
 
-from dfvfs.lib import definitions as dfvfs_definitions
-
 # The following import makes sure the output modules are registered.
 from plaso import output  # pylint: disable=unused-import
 
 from plaso.cli import extraction_tool
-from plaso.cli import logger
-from plaso.cli import status_view
 from plaso.cli import tool_options
 from plaso.cli import views
 from plaso.cli.helpers import manager as helpers_manager
 from plaso.engine import engine
 from plaso.engine import knowledge_base
-from plaso.engine import single_process as single_extraction_engine
-from plaso.lib import definitions
 from plaso.lib import errors
 from plaso.lib import loggers
-from plaso.multi_processing import extraction_engine as multi_extraction_engine
 from plaso.multi_processing import output_engine as multi_output_engine
 from plaso.parsers import manager as parsers_manager
 from plaso.storage import factory as storage_factory
@@ -83,11 +76,6 @@ class PstealTool(
       'And that is how you build a timeline using psteal...',
       '']))
 
-  _SOURCE_TYPES_TO_PREPROCESS = frozenset([
-      dfvfs_definitions.SOURCE_TYPE_DIRECTORY,
-      dfvfs_definitions.SOURCE_TYPE_STORAGE_MEDIA_DEVICE,
-      dfvfs_definitions.SOURCE_TYPE_STORAGE_MEDIA_IMAGE])
-
   def __init__(self, input_reader=None, output_writer=None):
     """Initializes the CLI tool object.
 
@@ -100,18 +88,13 @@ class PstealTool(
     super(PstealTool, self).__init__(
         input_reader=input_reader, output_writer=output_writer)
     self._artifacts_registry = None
-    self._command_line_arguments = None
     self._deduplicate_events = True
-    self._enable_sigsegv_handler = False
     self._knowledge_base = knowledge_base.KnowledgeBase()
     self._number_of_analysis_reports = 0
-    self._number_of_extraction_workers = 0
     self._output_format = None
     self._parsers_manager = parsers_manager.ParsersManager
     self._preferred_language = 'en-US'
     self._preferred_year = None
-    self._status_view_mode = status_view.StatusView.MODE_WINDOW
-    self._status_view = status_view.StatusView(self._output_writer, self.NAME)
     self._time_slice = None
     self._use_time_slicer = False
 
@@ -243,121 +226,6 @@ class PstealTool(
 
     self._output_writer.Write('Storage file is {0:s}\n'.format(
         self._storage_file_path))
-
-  def ExtractEventsFromSources(self):
-    """Processes the sources and extract events.
-
-    This is a stripped down copy of tools/log2timeline.py that doesn't support
-    the full set of flags. The defaults for these are hard coded in the
-    constructor of this class.
-
-    Raises:
-      BadConfigOption: if the storage file path is invalid or the storage
-          format not supported or an invalid collection filter was specified.
-      SourceScannerError: if the source scanner could not find a supported
-          file system.
-      UserAbort: if the user initiated an abort.
-    """
-    self._CheckStorageFile(self._storage_file_path, warn_about_existing=True)
-
-    self.ScanSource(self._source_path)
-
-    is_archive = False
-    if self._source_type == dfvfs_definitions.SOURCE_TYPE_FILE:
-      is_archive = self._IsArchiveFile(self._source_path_specs[0])
-      if is_archive:
-        self._source_type = definitions.SOURCE_TYPE_ARCHIVE
-
-    self._status_view.SetMode(self._status_view_mode)
-    self._status_view.SetSourceInformation(
-        self._source_path, self._source_type,
-        artifact_filters=self._artifact_filters,
-        filter_file=self._filter_file)
-
-    status_update_callback = (
-        self._status_view.GetExtractionStatusUpdateCallback())
-
-    self._output_writer.Write('\n')
-    self._status_view.PrintExtractionStatusHeader(None)
-    self._output_writer.Write('Processing started.\n')
-
-    session = engine.BaseEngine.CreateSession(
-        artifact_filter_names=self._artifact_filters,
-        command_line_arguments=self._command_line_arguments,
-        filter_file_path=self._filter_file,
-        preferred_encoding=self.preferred_encoding,
-        preferred_time_zone=self._preferred_time_zone,
-        preferred_year=self._preferred_year)
-
-    storage_writer = storage_factory.StorageFactory.CreateStorageWriter(
-        self._storage_format, session, self._storage_file_path)
-    if not storage_writer:
-      raise errors.BadConfigOption(
-          'Unsupported storage format: {0:s}'.format(self._storage_format))
-
-    single_process_mode = self._single_process_mode
-    if self._source_type == dfvfs_definitions.SOURCE_TYPE_FILE:
-      if not self._process_archives or not is_archive:
-        single_process_mode = True
-
-    if single_process_mode:
-      extraction_engine = single_extraction_engine.SingleProcessEngine()
-    else:
-      extraction_engine = multi_extraction_engine.ExtractionMultiProcessEngine(
-          number_of_worker_processes=self._number_of_extraction_workers,
-          worker_memory_limit=self._worker_memory_limit,
-          worker_timeout=self._worker_timeout)
-
-    # If the source is a directory or a storage media image
-    # run pre-processing.
-    if self._source_type in self._SOURCE_TYPES_TO_PREPROCESS:
-      self._PreprocessSources(extraction_engine, storage_writer)
-
-    configuration = self._CreateProcessingConfiguration(
-        extraction_engine.knowledge_base)
-
-    session.enabled_parser_names = (
-        configuration.parser_filter_expression.split(','))
-    session.parser_filter_expression = self._parser_filter_expression
-
-    self._SetExtractionPreferredTimeZone(extraction_engine.knowledge_base)
-
-    # TODO: set mount path in knowledge base with
-    # extraction_engine.knowledge_base.SetMountPath()
-    extraction_engine.knowledge_base.SetTextPrepend(self._text_prepend)
-
-    try:
-      extraction_engine.BuildCollectionFilters(
-          self._artifact_definitions_path, self._custom_artifacts_path,
-          extraction_engine.knowledge_base, self._artifact_filters,
-          self._filter_file)
-    except errors.InvalidFilter as exception:
-      raise errors.BadConfigOption(
-          'Unable to build collection filters with error: {0!s}'.format(
-              exception))
-
-    processing_status = None
-    if single_process_mode:
-      logger.debug('Starting extraction in single process mode.')
-
-      processing_status = extraction_engine.ProcessSources(
-          session, self._source_path_specs, storage_writer,
-          self._resolver_context, configuration,
-          status_update_callback=status_update_callback)
-
-    else:
-      logger.debug('Starting extraction in multi process mode.')
-
-      # The following overrides are needed because pylint 2.6.0 gets confused
-      # about which ProcessSources to check against.
-      # pylint: disable=no-value-for-parameter,unexpected-keyword-arg
-      processing_status = extraction_engine.ProcessSources(
-          session, self._source_path_specs, storage_writer, configuration,
-          enable_sigsegv_handler=self._enable_sigsegv_handler,
-          status_update_callback=status_update_callback,
-          storage_file_path=self._storage_file_path)
-
-    self._status_view.PrintExtractionSummary(processing_status)
 
   def ParseArguments(self, arguments):
     """Parses the command line arguments.
